+++ backend/assignment/serializers.py
@@ -6,4 +6,4 @@ from .models import Assignment
 class AssignmentSerializer(serializers.ModelSerializer):
     class Meta:
         model = Assignment
-        fields = '__all__'
+        fields = ('__all__')
